+++ packages/sdk/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ### Unreleased
 
+- add `unzip` to utils
 - add default indexing throttle to SDK to avoid subgraph rate limits
 - draw start block from config in wallet sdk
 - wallet sdk holds ref to `NocturneConfig`